affichage pos et angle bien

+++ src/parse_cub.c
@@ -6,7 +6,7 @@
 /*   By: mavileo <mavileo@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/17 03:09:19 by mavileo           #+#    #+#             */
-/*   Updated: 2020/01/15 13:40:28 by mavileo          ###   ########.fr       */
+/*   Updated: 2020/01/15 14:52:03 by mavileo          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -59,7 +59,7 @@ int		parse_cub(int fd, t_stru *stru)
 		map = strjoin_free_nl(map, line);
 	if ((tab_to_matrix(stru, map)) || check_stru(stru) || check_map(stru))
 		return (error_parsing());
-	stru->inter_rays = FOV / stru->res.x;
+	stru->inter_rays = (double)FOV / (double)stru->res.x;
 	//begin_pos(stru);
 	return (0);
 }
